src/main/java/com/slytechs/jnet/jnetruntime/bpf/vm/core/DefaultBpfExtension.java
package com.slytechs.jnet.jnetruntime.bpf.vm.core;

import java.nio.ByteBuffer;
import java.util.HashMap;
import java.util.Map;

import com.slytechs.jnet.jnetruntime.bpf.vm.api.BpfExtension;
import com.slytechs.jnet.jnetruntime.bpf.vm.api.ExecutionContext;
import com.slytechs.jnet.jnetruntime.bpf.vm.api.ExtensionContext;
import com.slytechs.jnet.jnetruntime.bpf.vm.api.ExtensionException;
import com.slytechs.jnet.jnetruntime.bpf.vm.api.ExtensionExecutionException;
import com.slytechs.jnet.jnetruntime.bpf.vm.api.ProtocolInfo;

/**
 * Default implementation of BPF extension providing basic functionality for
 * network protocol analysis and checksum verification.
 */
public class DefaultBpfExtension implements BpfExtension {

	private static final String NAME = "Default BPF Extension";
	private static final String VERSION = "1.0.0";

	// Extension opcodes
	private static final int OP_CHECK_L3_CSUM = 0x80;
	private static final int OP_CHECK_L4_CSUM = 0x81;
	private static final int OP_CHECK_TRUNCATED = 0x82;
	private static final int OP_CHECK_PROTO_OFFSET = 0x83;
	private static final int OP_GET_PROTO_FIELD = 0x84;
	private static final int OP_CHECK_PROTO_FLAGS = 0x85;

	// Error types
	private static final int ERROR_CHECKSUM = 1;
	private static final int ERROR_TRUNCATED = 2;
	private static final int ERROR_INVALID_PROTO = 3;

	private Map<String, String> config;
	private boolean initialized;

	@Override
	public String getName() {
		return NAME;
	}

	@Override
	public String getVersion() {
		return VERSION;
	}

	@Override
	public void initialize(ExtensionContext context) throws ExtensionException {
		if (initialized) {
			throw new ExtensionException("Extension already initialized");
		}

		this.config = new HashMap<>(context.getConfiguration());

		// Register extension opcodes
		context.registerOpcode(OP_CHECK_L3_CSUM, "CHECK_L3_CSUM");
		context.registerOpcode(OP_CHECK_L4_CSUM, "CHECK_L4_CSUM");
		context.registerOpcode(OP_CHECK_TRUNCATED, "CHECK_TRUNCATED");
		context.registerOpcode(OP_CHECK_PROTO_OFFSET, "CHECK_PROTO_OFFSET");
		context.registerOpcode(OP_GET_PROTO_FIELD, "GET_PROTO_FIELD");
		context.registerOpcode(OP_CHECK_PROTO_FLAGS, "CHECK_PROTO_FLAGS");

		initialized = true;
	}

	@Override
	public boolean executeInstruction(int opcode, int immediate,
			ExecutionContext context) throws ExtensionException {
		if (!initialized) {
			throw new ExtensionException("Extension not initialized");
		}

		try {
			switch (opcode) {
			case OP_CHECK_L3_CSUM:
				return checkLayer3Checksum(immediate, context);

			case OP_CHECK_L4_CSUM:
				return checkLayer4Checksum(immediate, context);

			case OP_CHECK_TRUNCATED:
				return checkTruncated(immediate, context);

			case OP_CHECK_PROTO_OFFSET:
				return checkProtocolOffset(immediate, context);

			case OP_GET_PROTO_FIELD:
				return getProtocolField(immediate, context);

			case OP_CHECK_PROTO_FLAGS:
				return checkProtocolFlags(immediate, context);

			default:
				return false; // Opcode not handled
			}
		} catch (Exception e) {
			throw new ExtensionExecutionException(
					"Error executing instruction: " + e.getMessage());
		}
	}

	@Override
	public void cleanup() {
		initialized = false;
		config.clear();
	}

	/**
	 * Checks Layer 3 checksum.
	 */
	private boolean checkLayer3Checksum(int immediate, ExecutionContext context) {
		ProtocolInfo l3Proto = context.getProtocolInfo(3);
		if (l3Proto == null) {
			context.setError(ERROR_INVALID_PROTO, 3);
			return true;
		}

		ByteBuffer packet = context.getPacketBuffer();
		int offset = l3Proto.getOffset();
		int length = l3Proto.getLength();

		// Verify we have enough data
		if (offset + length > packet.limit()) {
			context.setError(ERROR_TRUNCATED, length);
			return true;
		}

		// Calculate and verify checksum based on protocol type
		boolean valid = false;
		switch (l3Proto.getType()) {
		case 0x0800: // IPv4
			valid = verifyIPv4Checksum(packet, offset, length);
			break;
		default:
			// No checksum verification for other protocols
			valid = true;
		}

		if (!valid) {
			context.setError(ERROR_CHECKSUM, 3);
		}

		return true;
	}

	/**
	 * Checks Layer 4 checksum.
	 */
	private boolean checkLayer4Checksum(int immediate, ExecutionContext context) {
		ProtocolInfo l4Proto = context.getProtocolInfo(4);
		if (l4Proto == null) {
			context.setError(ERROR_INVALID_PROTO, 4);
			return true;
		}

		ByteBuffer packet = context.getPacketBuffer();
		int offset = l4Proto.getOffset();
		int length = l4Proto.getLength();

		// Verify we have enough data
		if (offset + length > packet.limit()) {
			context.setError(ERROR_TRUNCATED, length);
			return true;
		}

		// Calculate and verify checksum based on protocol type
		boolean valid = false;
		switch (l4Proto.getType()) {
		case 6: // TCP
			valid = verifyTCPChecksum(packet, offset, length, context);
			break;
		case 17: // UDP
			valid = verifyUDPChecksum(packet, offset, length, context);
			break;
		default:
			// No checksum verification for other protocols
			valid = true;
		}

		if (!valid) {
			context.setError(ERROR_CHECKSUM, 4);
		}

		return true;
	}

	/**
	 * Checks if packet is truncated.
	 */
	private boolean checkTruncated(int immediate, ExecutionContext context) {
		ByteBuffer packet = context.getPacketBuffer();
		if (immediate > packet.limit()) {
			context.setError(ERROR_TRUNCATED, immediate);
		}
		return true;
	}

	/**
	 * Checks protocol offset.
	 */
	private boolean checkProtocolOffset(int immediate, ExecutionContext context) {
		int layer = immediate & 0xFF;
		int expectedOffset = immediate >>> 8;

		ProtocolInfo proto = context.getProtocolInfo(layer);
		if (proto == null || proto.getOffset() != expectedOffset) {
			context.setError(ERROR_INVALID_PROTO, layer);
		}

		return true;
	}

	/**
	 * Gets protocol field value.
	 */
	private boolean getProtocolField(int immediate, ExecutionContext context) {
		int layer = immediate & 0xFF;
		int field = immediate >>> 8;

		ProtocolInfo proto = context.getProtocolInfo(layer);
		if (proto == null || !proto.hasField(field)) {
			context.setError(ERROR_INVALID_PROTO, layer);
			return true;
		}

		context.setRegister(0, proto.getField(field));
		return true;
	}

	/**
	 * Checks protocol flags.
	 */
	private boolean checkProtocolFlags(int immediate, ExecutionContext context) {
		int layer = immediate & 0xFF;
		int flags = immediate >>> 8;

		ProtocolInfo proto = context.getProtocolInfo(layer);
		if (proto == null) {
			context.setError(ERROR_INVALID_PROTO, layer);
			return true;
		}

		if ((proto.getFlags() & flags) != flags) {
			context.setRegister(0, 0);
		} else {
			context.setRegister(0, 1);
		}

		return true;
	}

	/**
	 * Verifies IPv4 header checksum.
	 */
	private boolean verifyIPv4Checksum(ByteBuffer packet, int offset, int length) {
		// Implementation would go here
		// This is a placeholder - real implementation would calculate
		// the actual IPv4 header checksum
		return true;
	}

	/**
	 * Verifies TCP checksum.
	 */
	private boolean verifyTCPChecksum(ByteBuffer packet, int offset, int length,
			ExecutionContext context) {
		// Implementation would go here
		// This is a placeholder - real implementation would calculate
		// the actual TCP checksum including pseudo-header
		return true;
	}

	/**
	 * Verifies UDP checksum.
	 */
	private boolean verifyUDPChecksum(ByteBuffer packet, int offset, int length,
			ExecutionContext context) {
		// Implementation would go here
		// This is a placeholder - real implementation would calculate
		// the actual UDP checksum including pseudo-header
		return true;
	}
}